# documentation/api/planning-core.openapi.yaml
openapi: 3.1.0
info:
  title: ConnectSphere planning-core API
  version: 0.2.0
  summary: The staff-facing API (identity and event modules) as built in Sprint 1 and EN-01.
  description: |
    planning-core is one deployable holding the identity, event, venue, equipment and
    change modules (ADR-0004). This file describes the routes that exist today; venue,
    equipment and change routes are added here, in the same pull request as their code
    (ADR-0015).

    **Conventions** (`implementation.md` §5):
    - Every refusal uses the error envelope below. `message` is written for the user.
      `fields[]` names every field that failed, not only the first.
    - A resource outside the caller's access scope (A3) is a 404, never a filtered 200,
      so its existence is not revealed.
    - Lists return `{ items, nextCursor }`. `nextCursor` is always null until pagination
      lands.
    - Every response echoes `X-Correlation-Id`. Send one to tie your logs to ours.
    - Timestamps are RFC 3339 UTC strings.

    `tests/api/openapiRoutes.unit.test.ts` (planning-core) fails CI if a route exists in
    the code but not here, or here but not in the code.

    In local development the Vite proxy serves identity routes under `/identity` and
    event routes under `/event`. Both prefixes reach the same planning-core process.
  contact:
    name: ConnectSphere team (IS212 G1)
  license:
    name: Proprietary (IS212 coursework)
    identifier: LicenseRef-IS212-coursework
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/equipment/types:
    get:
      tags: [Equipment]
      summary: List equipment inventory (P2)
      description: Event Coordinators and Technical Support Staff. Total quantity is stored for bulk stock and derived from unit records for serialized equipment. P1's availability check picks its equipment type from this list.
      operationId: listEquipmentTypes
      responses:
        "200":
          description: Equipment types in name order.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/EquipmentTypeListItem" }
                  nextCursor: { type: [string, "null"] }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
    post:
      tags: [Equipment]
      summary: Create an equipment type (P2)
      description: |
        Technical Support Staff only. Bulk types carry a non-negative quantity;
        serialized types carry unique unit labels. The creator, time, and initial
        quantity are recorded in inventory history in the same transaction.
      operationId: createEquipmentType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipmentTypeInput" }
      responses:
        "201":
          description: Created, including individually tracked units where applicable.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentType" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
  /api/v1/equipment/types/{id}:
    parameters:
      - $ref: "#/components/parameters/RecordId"
    get:
      tags: [Equipment]
      summary: Open an equipment type (P2)
      operationId: getEquipmentType
      responses:
        "200":
          description: Type details, units and active or removed unavailability records.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentType" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
    put:
      tags: [Equipment]
      summary: Update an equipment type and its held quantity (P2)
      description: |
        Technical Support Staff only. A reduction that would remove stock reserved
        during an overlapping period is refused without changing inventory; the
        refusal names each affected event reference and reserved quantity. Serialized
        stock is changed by synchronising its unit labels.
      operationId: updateEquipmentType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipmentTypeInput" }
      responses:
        "200":
          description: Updated; inventory history records actor, time and previous/new quantity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentType" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`INSUFFICIENT_EQUIPMENT`: reducing held stock would invalidate reservations; the message names affected events and quantities."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "422":
          description: "`VALIDATION_FAILED`: an existing type's tracking method cannot be changed."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
  /api/v1/equipment/types/{id}/unavailability:
    parameters:
      - $ref: "#/components/parameters/RecordId"
    post:
      tags: [Equipment]
      summary: Mark equipment unavailable for a period (P2)
      description: |
        Technical Support Staff only. Selects one unit for serialized equipment or a
        positive quantity for bulk stock. The reason is mandatory and the end must be
        later than the start. The change and its previous/new available quantity are audited.
      operationId: createEquipmentUnavailability
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EquipmentUnavailabilityInput" }
      responses:
        "201":
          description: The equipment is out of service for that half-open period.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentUnavailability" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: "`VALIDATION_FAILED` for a unit outside the type, or `INSUFFICIENT_EQUIPMENT` when the unavailable bulk quantity exceeds stock held."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
  /api/v1/equipment/types/{id}/availability:
    parameters:
      - $ref: "#/components/parameters/RecordId"
    get:
      tags: [Equipment]
      summary: Check numeric availability and shortfall over a period (P1)
      description: |
        Technical Support Staff only. Holds the equipment-type row lock within one
        transaction and writes no inventory, reservations, history or outbox data.
        Bulk availability is total minus peak concurrent combined reserved and
        unavailable quantities inside the half-open period (CQ-02). Individually
        tracked units must be free for the whole period; a unit both reserved and
        unavailable is excluded once. Touching periods do not overlap. Availability
        is clamped to zero. A shortage is a successful check, not a reservation refusal.
      operationId: checkEquipmentAvailability
      parameters:
        - name: startsAt
          in: query
          required: true
          description: RFC 3339 instant with an explicit timezone offset.
          schema: { type: string, format: date-time }
        - name: endsAt
          in: query
          required: true
          description: RFC 3339 instant later than startsAt, with an explicit timezone offset.
          schema: { type: string, format: date-time }
        - name: requestedQuantity
          in: query
          required: true
          description: Whole quantity to compare against availability, including zero.
          schema: { type: integer, minimum: 0, maximum: 9007199254740991 }
      responses:
        "200":
          description: Numeric availability, requested quantity and shortfall. No reservation is created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EquipmentAvailabilityResult" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimitExceeded" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/IdentityUnavailable" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A Supabase Auth access token. Every module verifies it itself.
  parameters:
    RecordId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    RateLimitExceeded:
      description: "`RATE_LIMITED`: 120 equipment requests per minute per client IP and process, shared across this router's endpoints. Enforced before authentication. Counters reset on process restart."
      headers:
        Retry-After:
          description: Seconds until this client's request window resets.
          schema: { type: integer, minimum: 1 }
        RateLimit:
          description: Combined limit, remaining and reset values (draft-7).
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    ValidationFailed:
      description: "`VALIDATION_FAILED`: `fields[]` names every field that failed."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthenticated:
      description: "`UNAUTHENTICATED`: no bearer token, or an invalid or expired one."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Forbidden:
      description: "`ROLE_NOT_AUTHORISED` (A2), or `NO_ROLE_ASSIGNED`."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: "`EVENT_NOT_FOUND`, `DRAFT_NOT_FOUND`, `REASSIGNMENT_PROPOSAL_NOT_FOUND` or `VENUE_NOT_FOUND` or `EQUIPMENT_TYPE_NOT_FOUND`. Also returned when the record exists but is outside the caller's scope (A3)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    IdentityUnavailable:
      description: "`IDENTITY_UNAVAILABLE`: the caller's role could not be read, so nothing was changed (CP: refuse rather than guess)."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    ErrorCode:
      type: string
      description: The shared list in `backend/packages/contracts/src/errorCodes.ts`.
      enum:
        - UNAUTHENTICATED
        - NO_ROLE_ASSIGNED
        - UNKNOWN_RESOURCE
        - INVALID_CREDENTIALS
        - ACCOUNT_DEACTIVATED
        - VALIDATION_FAILED
        - ROLE_NOT_AUTHORISED
        - EVENT_NOT_FOUND
        - DRAFT_NOT_FOUND
        - STATUS_TRANSITION_NOT_PERMITTED
        - EVENT_ALREADY_DECIDED # no longer emitted since F1 (decisions refuse with STATUS_TRANSITION_NOT_PERMITTED)
        - DRAFT_ALREADY_SUBMITTED
        - NO_OPEN_CLARIFICATION
        - IDENTITY_UNAVAILABLE
        - REASSIGNMENT_NOT_PERMITTED
        - NOMINEE_NOT_ELIGIBLE
        - SELF_NOMINATION_NOT_PERMITTED
        - REASSIGNMENT_ALREADY_PENDING
        - REASSIGNMENT_PROPOSAL_NOT_FOUND
        - VENUE_NOT_FOUND
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code: { $ref: "#/components/schemas/ErrorCode" }
            message: { type: string, description: Written for the user; names the specific values. }
            details: { type: object, additionalProperties: true }
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
            correlationId: { type: [string, "null"] }
...
    EquipmentTypeInput:
      type: object
      required: [name, description, characteristics, kind, totalQuantity, unitLabels]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        characteristics:
          type: object
          description: Technical characteristic names and their values.
          additionalProperties: { type: string }
        kind: { type: string, enum: [BULK, SERIALIZED] }
        totalQuantity:
          type: [integer, "null"]
          minimum: 0
          description: Required for BULK and null for SERIALIZED.
        unitLabels:
          type: array
          uniqueItems: true
          description: Empty for BULK; the unique physical-unit labels for SERIALIZED.
          items: { type: string, minLength: 1 }
    EquipmentUnit:
      type: object
      required: [id, label]
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
    EquipmentUnavailabilityInput:
      type: object
      required: [unitId, quantity, startsAt, endsAt, reason]
      properties:
        unitId:
          type: [string, "null"]
          format: uuid
          description: Required for SERIALIZED and null for BULK.
        quantity:
          type: [integer, "null"]
          minimum: 1
          description: Required for BULK and null for SERIALIZED.
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time, description: Must be later than startsAt. }
        reason: { type: string, minLength: 1 }
    EquipmentUnavailability:
      allOf:
        - $ref: "#/components/schemas/EquipmentUnavailabilityInput"
        - type: object
          required: [id, equipmentTypeId, status, createdAt, createdBy, updatedAt, updatedBy]
          properties:
            id: { type: string, format: uuid }
            equipmentTypeId: { type: string, format: uuid }
            status: { type: string, enum: [ACTIVE, REMOVED] }
            createdAt: { type: string, format: date-time }
            createdBy: { type: [string, "null"], format: uuid }
            updatedAt: { type: string, format: date-time }
            updatedBy: { type: [string, "null"], format: uuid }
    EquipmentType:
      allOf:
        - $ref: "#/components/schemas/EquipmentTypeInput"
        - type: object
          required: [id, units, unavailability, history, createdAt, createdBy, updatedAt, updatedBy]
          properties:
            id: { type: string, format: uuid }
            units:
              type: array
              items: { $ref: "#/components/schemas/EquipmentUnit" }
            unavailability:
              type: array
              items: { $ref: "#/components/schemas/EquipmentUnavailability" }
            history:
              type: array
              description: Newest-first immutable audit evidence for every inventory and unavailability change.
              items: { $ref: "#/components/schemas/InventoryHistoryEntry" }
            createdAt: { type: string, format: date-time }
            createdBy: { type: [string, "null"], format: uuid }
            updatedAt: { type: string, format: date-time }
            updatedBy: { type: [string, "null"], format: uuid }
    EquipmentTypeListItem:
      type: object
      required: [id, name, description, kind, totalQuantity]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        kind: { type: string, enum: [BULK, SERIALIZED] }
        totalQuantity:
          type: integer
          minimum: 0
          description: Stored bulk quantity or count of serialized unit records.
    InventoryHistoryEntry:
      type: object
      required: [id, action, previousQuantity, newQuantity, changes, actorUserId, actorRole, occurredAt]
      properties:
        id: { type: string, format: uuid }
        action: { type: string, enum: [TYPE_CREATED, TYPE_UPDATED, UNAVAILABILITY_RECORDED] }
        previousQuantity: { type: integer, minimum: 0 }
        newQuantity: { type: integer, minimum: 0 }
        changes: { type: object, additionalProperties: true }
        actorUserId: { type: string, format: uuid }
        actorRole: { type: string, const: TECH_SUPPORT_STAFF }
        occurredAt: { type: string, format: date-time }
    EquipmentAvailabilityResult:
      type: object
      required: [equipmentTypeId, kind, totalQuantity, availableQuantity, requestedQuantity, shortfallQuantity, startsAt, endsAt]
      properties:
        equipmentTypeId: { type: string, format: uuid }
        kind: { type: string, enum: [BULK, SERIALIZED] }
        totalQuantity: { type: integer, minimum: 0 }
        availableQuantity: { type: integer, minimum: 0 }
        requestedQuantity: { type: integer, minimum: 0 }
        shortfallQuantity: { type: integer, minimum: 0 }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }